+++ app/src/main/java/org/simple/clinic/setup/SetupActivityUpdate.kt
@@ -32,6 +32,15 @@ class SetupActivityUpdate(
       is CountryAndDeploymentSaved -> dispatch(DeleteStoredCountryV1)
       is StoredCountryV1Deleted -> dispatch(GoToMainActivity)
       DatabaseEncryptionFinished -> dispatch(CheckIfAppCanRun)
+      is MinimumMemoryChecked -> minimumMemoryChecked(event)
+    }
+  }
+
+  private fun minimumMemoryChecked(event: MinimumMemoryChecked): Next<SetupActivityModel, SetupActivityEffect> {
+    return if (event.hasMinimumMemory) {
+      dispatch(ExecuteDatabaseEncryption)
+    } else {
+      dispatch(CheckIfAppCanRun)
     }
   }