+++ sharedTestCode/src/main/java/org/simple/sharedTestCode/util/TestMinimumMemoryChecker.kt
@@ -1,0 +1,16 @@
+package org.simple.sharedTestCode.util
+
+import org.simple.clinic.util.MinimumMemoryChecker
+
+class TestMinimumMemoryChecker : MinimumMemoryChecker {
+
+  private var hasMinimumMemory: Boolean = false
+
+  override fun hasMinimumRequiredMemory(): Boolean {
+    return hasMinimumMemory
+  }
+
+  fun update(hasMinimumMemory: Boolean) {
+    this.hasMinimumMemory = hasMinimumMemory
+  }
+}